reorganized pages into page folders fixed any errs
