+++ src/index.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 
 import store from './redux/store';
 
-import App from './components/App/App';
+import App from './pages/App/App';
 
 ReactDOM.render(
   <Provider store={store}>
