+++ src/pages/App/App.jsx
@@ -8,8 +8,8 @@ import {
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import Nav from '../Nav/Nav';
-import Footer from '../Footer/Footer';
+import Nav from '../../components/Nav/Nav';
+import Footer from '../../components/Footer/Footer';
 
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
