+++ src/pages/UserPage/UserPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LogOutButton from '../LogOutButton/LogOutButton';
+import LogOutButton from '../../components/LogOutButton/LogOutButton';
 import {useSelector} from 'react-redux';
 
 function UserPage() {
